+++ public/admin/_app/immutable/entry/start.CZBzUAXp.js
@@ -1,0 +1,1 @@
+import{n as e,s as t}from"../chunks/Du8AfVOj.js";export{t as load_css,e as start};