+++ public/admin/_app/immutable/nodes/43.491P8vZB.js
@@ -1,1 +1,0 @@
-import{Ht as e}from"../chunks/Dqbbk3fc.js";import{c as t}from"../chunks/Du8n6nmR.js";import{t as n}from"../chunks/BXZ4Q3d1.js";var r=e({load:()=>i}),i=()=>{throw n(308,`${t}/settings/users`)};export{r as universal};